+++ src/resources/js/Pages/Dashboard.tsx
@@ -1,13 +1,13 @@
 import { Head } from '@inertiajs/react';
 import AuthenticatedLayout from '@/Layouts/AuthenticatedLayout';
 import { PageProps } from '@/types';
-import ManagementButtonList from '@/Features/Management/Component/ManagementButtonList';
+import TimeManagement from '@/Features/Management/Component/TimeManagement';
 
 export default function Dashboard({ auth }: PageProps) {
   return (
     <AuthenticatedLayout user={auth.user}>
       <Head title="Dashboard" />
-      <ManagementButtonList />
+      <TimeManagement />
     </AuthenticatedLayout>
   );
 }
